+++ util-read.R
@@ -345,8 +345,8 @@ read.issues = function(data.path, issues.sources = c("jira", "github")) {
     issue.data[["closing.date"]] = get.date.from.string(issue.data[["closing.date"]])
 
     # fix all dates to be after the creation date.
-    # violations can happen for commit_added events if the commit was made before the PR was opened
-    # the original date for commit_added events is stored in event.info.2 in any case
+    # violations can happen for "commit_added events" if the commit was made before the PR was opened
+    # the original date for "commit_added" events is stored in "event.info.2" in any case
     commit.added.events = issue.data[["event.name"]] == "commit_added"
     issue.data[commit.added.events, "event.info.2"] = get.date.string(issue.data[commit.added.events, "date"])
     commit.added.events.before.creation = commit.added.events & (issue.data["date"] < issue.data["creation.date"])